make generic broadcast screen component
