+++ client/src/containers/convo/CreateConvo.js
@@ -17,20 +17,11 @@ import { addPages, removePage } from '../../redux/actions/pages.actions'
 import { Link } from 'react-router'
 import { checkConditions } from '../polls/utility'
 import { validateFields } from './utility'
-import Image from './Image'
-import List from './List'
-import Video from './Video'
-import Audio from './Audio'
-import File from './File'
-import Text from './Text'
-import Card from './Card'
-import Gallery from './Gallery'
 import Targeting from './Targeting'
-import Media from './Media'
+import GenericMessage from '../../components/GenericMessage'
 // import DragSortableList from 'react-drag-sortable'
 import AlertContainer from 'react-alert'
 import { ModalContainer, ModalDialog } from 'react-modal-dialog'
-import StickyDiv from 'react-stickydiv'
 import { getuserdetails, getFbAppId, getAdminSubscriptions } from '../../redux/actions/basicinfo.actions'
 import { registerAction } from '../../utility/socketio'
 import {loadTags} from '../../redux/actions/tags.actions'
@@ -71,23 +62,9 @@ class CreateConvo extends React.Component {
     props.loadTags()
     this.showGuideLinesDialog = this.showGuideLinesDialog.bind(this)
     this.closeGuideLinesDialog = this.closeGuideLinesDialog.bind(this)
-    this.showResetAlertDialog = this.showResetAlertDialog.bind(this)
-    this.closeResetAlertDialog = this.closeResetAlertDialog.bind(this)
     this.handleSendBroadcast = this.handleSendBroadcast.bind(this)
-    this.handleMedia = this.handleMedia.bind(this)
-    this.handleText = this.handleText.bind(this)
-    this.handleCard = this.handleCard.bind(this)
-    this.handleGallery = this.handleGallery.bind(this)
-    this.handleList = this.handleList.bind(this)
-    this.handleImage = this.handleImage.bind(this)
-    this.handleFile = this.handleFile.bind(this)
-    this.removeComponent = this.removeComponent.bind(this)
     this.sendConvo = this.sendConvo.bind(this)
     this.testConvo = this.testConvo.bind(this)
-    this.newConvo = this.newConvo.bind(this)
-    this.showDialog = this.showDialog.bind(this)
-    this.closeDialog = this.closeDialog.bind(this)
-    this.renameTitle = this.renameTitle.bind(this)
     this.goBack = this.goBack.bind(this)
     this.onNext = this.onNext.bind(this)
     this.onPrevious = this.onPrevious.bind(this)
@@ -98,6 +75,11 @@ class CreateConvo extends React.Component {
     this.closeInvalidSession = this.closeInvalidSession.bind(this)
     this.showGuideLinesImageDialog = this.showGuideLinesImageDialog.bind(this)
     this.closeGuideLinesImageDialog = this.closeGuideLinesImageDialog.bind(this)
+    this.handleChange = this.handleChange.bind(this)
+  }
+
+  handleChange (state) {
+    this.setState(state)
   }
 
   onNext (e) {
@@ -177,15 +159,15 @@ class CreateConvo extends React.Component {
     this.top.scrollIntoView({behavior: 'instant'})
   }
   componentDidMount () {
-    const hostname =  window.location.hostname;
-    let title = '';
-    if(hostname.includes('kiboengage.cloudkibo.com')) {
-      title = 'KiboEngage';
+    const hostname = window.location.hostname
+    let title = ''
+    if (hostname.includes('kiboengage.cloudkibo.com')) {
+      title = 'KiboEngage'
     } else if (hostname.includes('kibochat.cloudkibo.com')) {
-      title = 'KiboChat';
+      title = 'KiboChat'
     }
 
-    document.title = `${title} | Create Broadcast`;
+    document.title = `${title} | Create Broadcast`
     this.scrollToTop()
     this.initTab()
 
@@ -223,215 +205,12 @@ class CreateConvo extends React.Component {
   closeGuideLinesDialog () {
     this.setState({isShowingModalGuideLines: false})
   }
-  showResetAlertDialog () {
-    if (this.state.broadcast.length > 0 || this.state.list.length > 0) {
-      this.setState({isShowingModalResetAlert: true})
-    }
-  }
-
-  closeResetAlertDialog () {
-    this.setState({isShowingModalResetAlert: false})
-  }
-  showDialog () {
-    this.setState({isShowingModal: true})
-  }
-
-  closeDialog () {
-    this.setState({isShowingModal: false})
-  }
-
-  renameTitle () {
-    if (this.titleConvo.value === '') {
-      return
-    }
-    this.setState({convoTitle: this.titleConvo.value})
-    this.closeDialog()
-  }
 
   gotoView (event) {
     this.props.history.push({
       pathname: `/broadcasts`
 
     })
-  }
-
-  handleText (obj) {
-    console.log('handleText', obj)
-    var temp = this.state.broadcast
-    var isPresent = false
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i].text = obj.text
-        if (obj.button.length > 0) {
-          temp[i].buttons = obj.button
-        } else {
-          delete temp[i].buttons
-        }
-        isPresent = true
-      }
-    })
-
-    if (!isPresent) {
-      if (obj.button.length > 0) {
-        temp.push({id: obj.id, text: obj.text, componentType: 'text', buttons: obj.button})
-      } else {
-        temp.push({id: obj.id, text: obj.text, componentType: 'text'})
-      }
-    }
-
-    this.setState({broadcast: temp})
-  }
-
-  handleCard (obj) {
-    if (obj.error) {
-      if (obj.error === 'invalid image') {
-        this.msg.error('Please select an image of type jpg, gif, bmp or png')
-      }
-      return
-    }
-    var temp = this.state.broadcast
-    var isPresent = false
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i].fileName = obj.fileName
-        temp[i].fileurl = obj.fileurl
-        temp[i].image_url = obj.image_url
-        temp[i].size = obj.size
-        temp[i].type = obj.type
-        temp[i].title = obj.title
-        temp[i].buttons = obj.buttons
-        temp[i].description = obj.description
-        if (obj.default_action && obj.default_action !== '') {
-          temp[i].default_action = obj.default_action
-        } else if (temp[i].default_action) {
-          delete temp[i].default_action
-        }
-        isPresent = true
-      }
-    })
-    if (!isPresent) {
-      temp.push(obj)
-    }
-    this.setState({broadcast: temp})
-  }
-
-  handleMedia (obj) {
-    if (obj.error) {
-      if (obj.error === 'invalid image') {
-        this.msg.error('Please select an image of type jpg, gif, bmp or png')
-        return
-      }
-      if (obj.error === 'file size error') {
-        this.msg.error('File size cannot exceed 25MB')
-        return
-      }
-      if (obj.error === 'invalid file') {
-        this.msg.error('File is not valid')
-        return
-      }
-    }
-    var temp = this.state.broadcast
-    var isPresent = false
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i].fileName = obj.fileName
-        temp[i].mediaType = obj.mediaType
-        temp[i].fileurl = obj.fileurl
-        temp[i].image_url = obj.image_url
-        temp[i].size = obj.size
-        temp[i].type = obj.type
-        temp[i].buttons = obj.buttons
-        isPresent = true
-      }
-    })
-    if (!isPresent) {
-      temp.push(obj)
-    }
-    this.setState({broadcast: temp})
-  }
-
-  handleGallery (obj) {
-    var temp = this.state.broadcast
-    var isPresent = false
-    if (obj.cards) {
-      obj.cards.forEach((d) => {
-        delete d.id
-      })
-    }
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i].cards = obj.cards
-        isPresent = true
-      }
-    })
-    if (!isPresent) {
-      temp.push(obj)
-    }
-    this.setState({broadcast: temp})
-  }
-
-  handleImage (obj) {
-    var temp = this.state.broadcast
-    var isPresent = false
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i] = obj
-        isPresent = true
-      }
-    })
-
-    if (!isPresent) {
-      temp.push(obj)
-    }
-
-    this.setState({broadcast: temp})
-  }
-
-  handleFile (obj) {
-    var temp = this.state.broadcast
-    var isPresent = false
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i] = obj
-        isPresent = true
-      }
-    })
-
-    if (!isPresent) {
-      temp.push(obj)
-    }
-
-    this.setState({broadcast: temp})
-  }
-  handleList (obj) {
-    console.log('in create convo handleList', obj)
-    var temp = this.state.broadcast
-    var isPresent = false
-    obj.listItems.forEach((d) => {
-      delete d.id
-    })
-    temp.map((data, i) => {
-      if (data.id === obj.id) {
-        temp[i].listItems = obj.listItems
-        temp[i].topElementStyle = obj.topElementStyle
-        temp[i].buttons = obj.buttons
-        isPresent = true
-      }
-    })
-    if (!isPresent) {
-      temp.push(obj)
-    }
-    console.log('temp', temp)
-    this.setState({broadcast: temp})
-  }
-
-  removeComponent (obj) {
-    console.log('obj in removeComponent', obj)
-    var temp = this.state.list.filter((component) => { return (component.props.id !== obj.id) })
-    var temp2 = this.state.broadcast.filter((component) => { return (component.id !== obj.id) })
-    console.log('temp', temp)
-    console.log('temp2', temp2)
-    this.setState({list: temp, broadcast: temp2})
   }
 
   sendConvo () {
@@ -546,10 +325,6 @@ class CreateConvo extends React.Component {
     }
   }
 
-  newConvo () {
-    this.setState({broadcast: [], list: []})
-  }
-
   goBack () {
     this.props.history.push({
       pathname: `/welcomeMessage`
@@ -592,6 +367,86 @@ class CreateConvo extends React.Component {
             </ModalDialog>
           </ModalContainer>
         }
+        {
+          this.state.isShowingModalGuideLines &&
+          <ModalContainer style={{width: '500px'}}
+            onClose={this.closeGuideLinesDialog}>
+            <ModalDialog style={{width: '500px'}}
+              onClose={this.closeGuideLinesDialog}>
+              <h4>Message Types</h4>
+              <p> Following are the types of messages that can be sent to facebook messenger.</p>
+              <div className='panel-group accordion' id='accordion1'>
+                <div className='panel panel-default'>
+                  <div className='panel-heading guidelines-heading'>
+                    <h4 className='panel-title'>
+                      <a className='guidelines-link accordion-toggle accordion-toggle-styled collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_1' aria-expanded='false'>Subscription Messages</a>
+                    </h4>
+                  </div>
+                  <div id='collapse_1' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
+                    <div className='panel-body'>
+                      <p>Subscription messages can&#39;t contain ads or promotional materials, but can be sent at any time regardless of time passed since last user activity. In order to send Subscription Messages, please apply for Subscription Messages Permission by following the steps given on this&nbsp;
+                      <a href='https://kibopush.com/subscription-messaging/' target='_blank'>link.</a></p>
+                    </div>
+                  </div>
+                </div>
+                <div className='panel panel-default'>
+                  <div className='panel-heading guidelines-heading'>
+                    <h4 className='panel-title'>
+                      <a className='guidelines-link accordion-toggle collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_2' aria-expanded='false'>Promotional Messages</a>
+                    </h4>
+                  </div>
+                  <div id='collapse_2' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
+                    <div className='panel-body'>
+                      Promotional messages can contain ads and promotional materials, but can only be sent to subscribers who were active in the past 24 hours.
+                    </div>
+                  </div>
+                </div>
+                <div className='panel panel-default'>
+                  <div className='panel-heading guidelines-heading'>
+                    <h4 className='panel-title'>
+                      <a className='guidelines-link accordion-toggle collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_3' aria-expanded='false'>Follow-Up Messages</a>
+                    </h4>
+                  </div>
+                  <div id='collapse_3' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
+                    <div className='panel-body'>
+                      After the end of the 24 hours window you have an ability to send "1 follow up message" to these recipients. After that you won&#39;t be able to send them ads or promotional messages until they interact with you again.
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </ModalDialog>
+          </ModalContainer>
+        }
+        {
+          this.state.isshowGuideLinesImageDialog &&
+          <ModalContainer style={{width: '500px'}}
+            onClose={this.closeGuideLinesImageDialog}>
+            <ModalDialog style={{width: '500px'}}
+              onClose={this.closeGuideLinesImageDialog}>
+              <p>Use the correct aspect ratio for your image. Photos in the generic template that aren't <strong>1.91:1 </strong>will be scaled or cropped. </p>
+
+            </ModalDialog>
+          </ModalContainer>
+        }
+
+        {
+          this.state.showMessengerModal &&
+          <ModalContainer style={{width: '500px'}}
+            onClick={() => { this.setState({showMessengerModal: false}) }}
+            onClose={() => { this.setState({showMessengerModal: false}) }}>
+            <ModalDialog style={{width: '500px'}}
+              onClick={() => { this.setState({showMessengerModal: false}) }}
+              onClose={() => { this.setState({showMessengerModal: false}) }}>
+              <h3 onClick={() => { this.setState({showMessengerModal: false}) }} >Connect to Messenger:</h3>
+              <MessengerPlugin
+                appId={this.props.fbAppId}
+                pageId={JSON.stringify(this.state.pageValue[0])}
+                passthroughParams={this.props.user._id}
+                onClick={() => { this.setState({showMessengerModal: false}) }}
+              />
+            </ModalDialog>
+          </ModalContainer>
+        }
         <div className='m-content'>
           <div className='m-alert m-alert--icon m-alert--air m-alert--square alert alert-dismissible m--margin-bottom-30' role='alert'>
             <div className='m-alert__icon'>
@@ -601,7 +456,7 @@ class CreateConvo extends React.Component {
               Need help in understanding how to create broadcasts? Here is the <a href='http://kibopush.com/broadcasts/' target='_blank'>documentation</a>.
               <br />
               View Facebook guidelines regarding types of messages here: <Link className='linkMessageTypes' style={{color: '#5867dd', cursor: 'pointer'}} onClick={this.showGuideLinesDialog} >Message Types</Link>
-          
+
               &ensp; and image guidelines: <Link className='linkMessageTypes' style={{color: '#5867dd', cursor: 'pointer'}} onClick={this.showGuideLinesImageDialog} >click here</Link>
             </div>
           </div>
@@ -629,7 +484,7 @@ class CreateConvo extends React.Component {
                       {
                         this.state.tabActive === 'broadcast' && this.props.location.state.module === 'convo' &&
                         <div className='pull-right'>
-                          <button className='btn btn-primary' style={{marginRight: '10px'}} onClick={this.showResetAlertDialog}>
+                          <button className='btn btn-primary' style={{marginRight: '10px'}} onClick={this.reset}>
                             Reset
                           </button>
                           <button className='btn btn-primary' disabled={(this.state.broadcast.length === 0)} onClick={this.onNext}>
@@ -671,220 +526,7 @@ class CreateConvo extends React.Component {
                       </ul>
                       <div className='tab-content'>
                         <div className='tab-pane fade active in' id='tab_1'>
-                          <div className='row'>
-                            <div className='col-lg-6 col-md-6 col-sm-12 col-xs-12'>
-                              <div className='row' >
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' id='text' onClick={() => { var temp = this.state.list; this.msg.info('New Text Component Added'); this.setState({list: [...temp, <Text id={timeStamp} pageId={this.state.pageId} key={timeStamp} handleText={this.handleText} onRemove={this.removeComponent} removeState />]}); this.handleText({id: timeStamp, text: '', button: []}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/text.png' alt='Text' style={{maxHeight: 25}} />
-                                      <h6>Text</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Image Component Added'); this.setState({list: [...temp, <Image id={timeStamp} pages={this.props.location.state.pages} key={timeStamp} handleImage={this.handleImage} onRemove={this.removeComponent} />]}); this.handleImage({id: timeStamp, componentType: 'image', image_url: '', fileurl: ''}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/picture.png' alt='Image' style={{maxHeight: 25}} />
-                                      <h6>Image</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Card Component Added'); this.setState({list: [...temp, <Card id={timeStamp} pageId={this.state.pageId} pages={this.props.location.state.pages} key={timeStamp} handleCard={this.handleCard} onRemove={this.removeComponent} singleCard />]}); this.handleCard({id: timeStamp, componentType: 'card', title: '', description: '', fileurl: '', buttons: []}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/card.png' alt='Card' style={{maxHeight: 25}} />
-                                      <h6>Card</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Gallery Component Added'); this.setState({list: [...temp, <Gallery id={timeStamp} pageId={this.state.pageId} pages={this.props.location.state.pages} key={timeStamp} handleGallery={this.handleGallery} onRemove={this.removeComponent} />]}); this.handleGallery({id: timeStamp, componentType: 'gallery', cards: []}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/layout.png' alt='Gallery' style={{maxHeight: 25}} />
-                                      <h6>Gallery</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className='row'>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Audio Component Added'); this.setState({list: [...temp, <Audio id={timeStamp} pages={this.props.location.state.pages} key={timeStamp} handleFile={this.handleFile} onRemove={this.removeComponent} />]}); this.handleFile({id: timeStamp, componentType: 'audio', fileurl: ''}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/speaker.png' alt='Audio' style={{maxHeight: 25}} />
-                                      <h6>Audio</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Video Component Added'); this.setState({list: [...temp, <Video id={timeStamp} pages={this.props.location.state.pages} key={timeStamp} handleFile={this.handleFile} onRemove={this.removeComponent} />]}); this.handleFile({id: timeStamp, componentType: 'video', fileurl: ''}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/video.png' alt='Video' style={{maxHeight: 25}} />
-                                      <h6>Video</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New File Component Added'); this.setState({list: [...temp, <File id={timeStamp} pages={this.props.location.state.pages} key={timeStamp} handleFile={this.handleFile} onRemove={this.removeComponent} />]}); this.handleFile({id: timeStamp, componentType: 'file', fileurl: ''}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/file.png' alt='File' style={{maxHeight: 25}} />
-                                      <h6>File</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New List Component Added'); this.setState({list: [...temp, <List id={timeStamp} pageId={this.state.pageId} pages={this.props.location.state.pages} key={timeStamp} handleList={this.handleList} onRemove={this.removeComponent} />]}); this.handleList({id: timeStamp, componentType: 'list', listItems: [], topElementStyle: 'compact'}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/list.png' alt='List' style={{maxHeight: 25}} />
-                                      <h6>List</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                              </div>
-                              <div className='row'>
-                                <div className='col-3'>
-                                  <div className='ui-block hoverbordercomponent' onClick={() => { var temp = this.state.list; this.msg.info('New Media Component Added'); this.setState({list: [...temp, <Media id={timeStamp} pageId={this.state.pageId} pages={this.props.location.state.pages} key={timeStamp} handleMedia={this.handleMedia} onRemove={this.removeComponent} />]}); this.handleMedia({id: timeStamp, componentType: 'media', fileurl: '', buttons: []}) }}>
-                                    <div className='align-center'>
-                                      <img src='https://cdn.cloudkibo.com/public/icons/media.png' alt='Media' style={{maxHeight: 25}} />
-                                      <h6>Media</h6>
-                                    </div>
-                                  </div>
-                                </div>
-                              </div>
-                            </div>
-                            <div className='col-lg-6 col-md-6 col-sm-12 col-xs-12'>
-                              <StickyDiv zIndex={1}>
-                                <div style={{border: '1px solid #ccc', borderRadius: '0px', backgroundColor: '#e1e3ea'}} className='ui-block'>
-                                  <div style={{padding: '5px'}}>
-                                    {this.props.location.state.module === 'welcome'
-                                    ? <h3>Welcome Message</h3>
-                                    : <h3>{this.state.convoTitle} <i onClick={this.showDialog} id='convoTitle' style={{cursor: 'pointer'}} className='fa fa-pencil-square-o' aria-hidden='true' /></h3>
-                                  }
-                                  </div>
-                                </div>
-                              </StickyDiv>
-                              {
-                                this.state.isShowingModal &&
-                                <ModalContainer style={{width: '500px'}}
-                                  onClose={this.closeDialog}>
-                                  <ModalDialog style={{width: '500px'}}
-                                    onClose={this.closeDialog}>
-                                    <h3>Rename:</h3>
-                                    <input style={{maxWidth: '300px', float: 'left', margin: 2}} ref={(c) => { this.titleConvo = c }} placeholder={this.state.convoTitle} type='text' className='form-control' />
-                                    <button style={{float: 'left', margin: 2}} onClick={this.renameTitle} className='btn btn-primary btn-sm' type='button'>Save</button>
-                                  </ModalDialog>
-                                </ModalContainer>
-                              }
-                              {
-                                this.state.isShowingModalResetAlert &&
-                                <ModalContainer style={{width: '500px'}}
-                                  onClose={this.closeResetAlertDialog}>
-                                  <ModalDialog style={{width: '500px'}}
-                                    onClose={this.closeResetAlertDialog}>
-                                    <p>Are you sure you want to reset the message ?</p>
-                                    <button style={{float: 'right', marginLeft: '10px'}}
-                                      className='btn btn-primary btn-sm'
-                                      onClick={() => {
-                                        this.newConvo()
-                                        this.closeResetAlertDialog()
-                                      }}>Yes
-                                    </button>
-                                    <button style={{float: 'right'}}
-                                      className='btn btn-primary btn-sm'
-                                      onClick={() => {
-                                        this.closeResetAlertDialog()
-                                      }}>Cancel
-                                    </button>
-                                  </ModalDialog>
-                                </ModalContainer>
-                              }
-                              {
-                                this.state.isShowingModalGuideLines &&
-                                <ModalContainer style={{width: '500px'}}
-                                  onClose={this.closeGuideLinesDialog}>
-                                  <ModalDialog style={{width: '500px'}}
-                                    onClose={this.closeGuideLinesDialog}>
-                                    <h4>Message Types</h4>
-                                    <p> Following are the types of messages that can be sent to facebook messenger.</p>
-                                    <div className='panel-group accordion' id='accordion1'>
-                                      <div className='panel panel-default'>
-                                        <div className='panel-heading guidelines-heading'>
-                                          <h4 className='panel-title'>
-                                            <a className='guidelines-link accordion-toggle accordion-toggle-styled collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_1' aria-expanded='false'>Subscription Messages</a>
-                                          </h4>
-                                        </div>
-                                        <div id='collapse_1' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
-                                          <div className='panel-body'>
-                                            <p>Subscription messages can&#39;t contain ads or promotional materials, but can be sent at any time regardless of time passed since last user activity. In order to send Subscription Messages, please apply for Subscription Messages Permission by following the steps given on this&nbsp;
-                                            <a href='https://kibopush.com/subscription-messaging/' target='_blank'>link.</a></p>
-                                          </div>
-                                        </div>
-                                      </div>
-                                      <div className='panel panel-default'>
-                                        <div className='panel-heading guidelines-heading'>
-                                          <h4 className='panel-title'>
-                                            <a className='guidelines-link accordion-toggle collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_2' aria-expanded='false'>Promotional Messages</a>
-                                          </h4>
-                                        </div>
-                                        <div id='collapse_2' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
-                                          <div className='panel-body'>
-                                            Promotional messages can contain ads and promotional materials, but can only be sent to subscribers who were active in the past 24 hours.
-                                          </div>
-                                        </div>
-                                      </div>
-                                      <div className='panel panel-default'>
-                                        <div className='panel-heading guidelines-heading'>
-                                          <h4 className='panel-title'>
-                                            <a className='guidelines-link accordion-toggle collapsed' data-toggle='collapse' data-parent='#accordion1' href='#collapse_3' aria-expanded='false'>Follow-Up Messages</a>
-                                          </h4>
-                                        </div>
-                                        <div id='collapse_3' className='panel-collapse collapse' aria-expanded='false' style={{height: '0px'}}>
-                                          <div className='panel-body'>
-                                            After the end of the 24 hours window you have an ability to send "1 follow up message" to these recipients. After that you won&#39;t be able to send them ads or promotional messages until they interact with you again.
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </div>
-                                  </ModalDialog>
-                                </ModalContainer>
-                              }
-                               {
-                                this.state.isshowGuideLinesImageDialog &&
-                                <ModalContainer style={{width: '500px'}}
-                                  onClose={this.closeGuideLinesImageDialog}>
-                                  <ModalDialog style={{width: '500px'}}
-                                    onClose={this.closeGuideLinesImageDialog}>
-                                   <p>Use the correct aspect ratio for your image. Photos in the generic template that aren't <strong>1.91:1 </strong>will be scaled or cropped. </p>
-
-                                  </ModalDialog>
-                                </ModalContainer>
-                              }
-
-                              {
-                                this.state.showMessengerModal &&
-                                <ModalContainer style={{width: '500px'}}
-                                  onClick={() => { this.setState({showMessengerModal: false}) }}
-                                  onClose={() => { this.setState({showMessengerModal: false}) }}>
-                                  <ModalDialog style={{width: '500px'}}
-                                    onClick={() => { this.setState({showMessengerModal: false}) }}
-                                    onClose={() => { this.setState({showMessengerModal: false}) }}>
-                                    <h3 onClick={() => { this.setState({showMessengerModal: false}) }} >Connect to Messenger:</h3>
-                                    <MessengerPlugin
-                                      appId={this.props.fbAppId}
-                                      pageId={JSON.stringify(this.state.pageId)}
-                                      passthroughParams={this.props.user._id}
-                                      onClick={() => { this.setState({showMessengerModal: false}) }}
-                                    />
-                                  </ModalDialog>
-                                </ModalContainer>
-                              }
-                              <div className='ui-block' style={{height: 90 + 'vh', overflowY: 'scroll', marginTop: '-15px', paddingLeft: 75, paddingRight: 75, paddingTop: 30, borderRadius: '0px', border: '1px solid #ccc'}}>
-                                {/* <h4  className="align-center" style={{color: '#FF5E3A', marginTop: 100}}> Add a component to get started </h4> */}
-                                {this.state.list}
-                              </div>
-
-                            </div>
-                          </div>
+                          <GenericMessage handleChange={this.handleChange} setReset={reset => { this.reset = reset }} convoTitle={this.state.convoTitle} titleEditable location={this.props.location} />
                         </div>
                         {
                           this.props.location.state.module === 'convo' &&
