+++ client/src/containers/convo/convo.js
@@ -108,17 +108,17 @@ class Convo extends React.Component {
     this.setState({isShowingModal: false})
   }
   componentDidMount () {
-    this.scrollToTop();
+    this.scrollToTop()
 
-    const hostname =  window.location.hostname;
-    let title = '';
-    if(hostname.includes('kiboengage.cloudkibo.com')) {
-      title = 'KiboEngage';
+    const hostname = window.location.hostname
+    let title = ''
+    if (hostname.includes('kiboengage.cloudkibo.com')) {
+      title = 'KiboEngage'
     } else if (hostname.includes('kibochat.cloudkibo.com')) {
-      title = 'KiboChat';
+      title = 'KiboChat'
     }
 
-    document.title = `${title} | Broadcast`;
+    document.title = `${title} | Broadcast`
   }
 
   displayData (n, broadcasts) {
